--- src/devices/Interop/Unix/Libc/Interop.close.cs
@@ -1,11 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Runtime.InteropServices;
-
-internal partial class Interop
-{
-    [DllImport(LibcLibrary)]
-    internal static extern int close(int fd);
-}
